+++ cli/proc_state.rs
@@ -1,5 +1,7 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
+use crate::args::import_map_from_text;
+use crate::args::import_map_from_value;
 use crate::args::CliOptions;
 use crate::args::DenoSubcommand;
 use crate::args::Flags;
@@ -43,8 +45,6 @@ use deno_core::futures;
 use deno_core::parking_lot::Mutex;
 use deno_core::parking_lot::RwLock;
 use deno_core::resolve_url_or_path;
-use deno_core::serde_json;
-use deno_core::url::Url;
 use deno_core::CompiledWasmModuleStore;
 use deno_core::ModuleSpecifier;
 use deno_core::SharedArrayBufferStore;
@@ -776,47 +776,6 @@ impl ProcState {
     }
 
     Ok(graph)
-  }
-}
-
-pub fn import_map_from_value(
-  specifier: &Url,
-  value: serde_json::Value,
-) -> Result<ImportMap, AnyError> {
-  debug_assert!(
-    !specifier.as_str().contains("../"),
-    "Import map specifier incorrectly contained ../: {}",
-    specifier.as_str()
-  );
-  let result = import_map::parse_from_value(specifier, value)?;
-  print_import_map_diagnostics(&result.diagnostics);
-  Ok(result.import_map)
-}
-
-pub fn import_map_from_text(
-  specifier: &Url,
-  json_text: &str,
-) -> Result<ImportMap, AnyError> {
-  debug_assert!(
-    !specifier.as_str().contains("../"),
-    "Import map specifier incorrectly contained ../: {}",
-    specifier.as_str()
-  );
-  let result = import_map::parse_from_json(specifier, json_text)?;
-  print_import_map_diagnostics(&result.diagnostics);
-  Ok(result.import_map)
-}
-
-fn print_import_map_diagnostics(diagnostics: &[String]) {
-  if !diagnostics.is_empty() {
-    warn!(
-      "Import map diagnostics:\n{}",
-      diagnostics
-        .iter()
-        .map(|d| format!("  - {}", d))
-        .collect::<Vec<_>>()
-        .join("\n")
-    );
   }
 }
